\name{xkcdcolors-package}
\alias{xkcdcolors-package}
\alias{xkcdcolors}
\docType{package}
\title{
\packageTitle{xkcdcolors}
}
\description{
\packageDescription{xkcdcolors}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{xkcdcolors}
\packageIndices{xkcdcolors}
}
\author{
\packageAuthor{xkcdcolors}

Maintainer: \packageMaintainer{xkcdcolors}
}
\references{
\url{http://blog.xkcd.com/2010/05/03/color-survey-results/}}


